Extract all Pydantic models and enums into scaffold/models.py
Consolidates the four data structure definitions (GoldenSuffix, Verdict,
OracleResult, PrefixResult) that were scattered across golden_suffixes.py,
oracle.py, and executor.py into a single shared models.py module. This
breaks the implicit coupling where modules had to import both data
definitions and logic from the same file, and prevents circular import
issues as the package grows.

- GoldenSuffix (from golden_suffixes.py): now imported by golden_suffixes
- Verdict + OracleResult (from oracle.py): now imported by oracle, executor
- PrefixResult (from executor.py): now imported by executor, runner

All four source files updated to import from scaffold.models instead of
defining inline or importing from peer modules. No behavioral changes.
